# Static compaction groups