--- a/db/system_keyspace.cc
+++ b/db/system_keyspace.cc
@@ -2304,8 +2304,13 @@ public:
                 static stats reduce(stats a, stats b) { return stats{a.total + b.total, a.free + b.free, a.entries + b.entries}; }
             };
             return map_reduce_tables<stats>([] (replica::table& t) {
-                const auto s = t.active_memtable().region().occupancy();
-                return stats{s.total_space(), s.free_space(), t.active_memtable().partition_count()};
+                logalloc::occupancy_stats s;
+                uint64_t partition_count = 0;
+                for (replica::memtable* active_memtable : t.active_memtables()) {
+                    s += active_memtable->region().occupancy();
+                    partition_count += active_memtable->partition_count();
+                }
+                return stats{s.total_space(), s.free_space(), partition_count};
             }, stats::reduce).then([] (stats s) {
                 return std::vector<std::pair<sstring, sstring>>{
                         {"memory_total", format("{}", s.total)},