--- a/replica/database.cc
+++ b/replica/database.cc
@@ -1238,6 +1238,7 @@ keyspace::make_column_family_config(const schema& s, const database& db) const {
     cfg.view_update_concurrency_semaphore = _config.view_update_concurrency_semaphore;
     cfg.view_update_concurrency_semaphore_limit = _config.view_update_concurrency_semaphore_limit;
     cfg.data_listeners = &db.data_listeners();
+    cfg.x_log2_compaction_groups = db_config.x_log2_compaction_groups();
 
     return cfg;
 }