… scylla-gdb.py
… test/boost/commitlog_test.cc
… test/boost/database_test.cc
… test/boost/memtable_test.cc
… test/boost/mutation_test.cc
… test/boost/sstable_test.hh
… test/perf/perf_fast_forward.cc