@bhalevy 2022-11-30
It's essential @tgrabiec reviews this too as it's highly relevant to tablets.