@avikivity 2022-12-01
> Allows static configuration of number of compaction groups per table per shard.
> 
> To bootstrap the project, config option x_log2_compaction_groups was added which controls both number of groups and partitioning within a shard.
> 
> With a value of 0 (default), it means 1 compaction group, therefore all tokens go there. With a value of 3, it means 8 compaction groups, and 3 most-significant-bits of tokens being used to decide which group owns the token. And so on.
> 
> It's still missing:
> 
> * integration with repair / streaming
> * integration with reshard / reshape.

We can go without these

> * automated testing

But let's have tests with the series.