@raphaelsc 2022-12-13
v2:
- Avoids lw_shared_ptr<compaction_group>
- table::compaction_groups() returns const ref
- table::active_memtables() returns std::vector<memtable*>
- don't reserve readers capacity in table::add_memtables_to_reader_list().
- introduce table::memtable_count() for memtable count across all groups
- point query will now pick memtables from a single compaction group
- use std::mem_fn() for reducing action on all groups to a single line, e.g.:
    std::ranges::any_of(compaction_groups(), std::mem_fn(&compaction_group::T));
- for_each_compaction_group(...) becomes parallel_foreach_compaction_group(...).
    we can operate on all groups in parallel
    if serialization is needed, that's done by lower layer like compaction_manager.
- table::do_trigger_compaction() is moved to compaction_group
- use 1 << _x_log2_compaction_groups when counting nr of groups.
- Reword description of option x_log2_compaction_groups
- Extend tests to stress multiple compaction groups