@scylladb-promoter 2022-12-14
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/3485/