@avikivity 2022-12-14
Looks good.

Please run the entire test suite (test.py) with multiple compaction groups.

Please adjust changelogs where you adapted a function with information about callers, and why the change is okay. For example, an function that is called from the api is fine to just accumulate, but rebuilding sstable sets after compaction is not okay to iterate over all compaction groups.