@raphaelsc 2022-12-14
> > Looks good.
> > Please run the entire test suite (test.py) with multiple compaction groups.
> 
> Some unit tests were written with a single compaction group in mind and they fail if I force they to run with multiple groups, e.g. they expect a certain # of sstables per shard, not one per group. I am thinking about incrementally enabling multiple groups until the entire suite is ready for it. database_test and mutation_test will already be stressing multiple groups, even with default settings in test.py. We can extend that to other relevant boost tests. For tests which launch a scylla instance, those can be easily forced to run with multiple groups with a simple patch to test.py (via a new option).
> 
> > Please adjust changelogs where you adapted a function with information about callers, and why the change is okay. For example, an function that is called from the api is fine to just accumulate, but rebuilding sstable sets after compaction is not okay to iterate over all compaction groups.

Good news is that only 8 out of 2607 tests failed, with multiple groups.