@raphaelsc 2022-12-19
v4: rebased