@avikivity 2022-12-20
So, the test suite now passes with multiple compaction groups (not talking about just the dedicated tests)?