@raphaelsc 2022-12-20
> So, the test suite now passes with multiple compaction groups (not talking about just the dedicated tests)?

@avikivity 7 tests out 2637 failed. most of them fail because they were written with a single group in mind, like expecting a certain # of sstables. I plan to address the failure after this series is merged. By the way, we need to run the suite with multiple groups on a regular base, I am planning to add a compilation option to override the default # of groups. I can also patch test.py to optionally launch scylla instances with multiple groups, but that won't cover boost and similar tests. For boost, we're running both single and multiple groups on a subset of tests and we can keep extending this subset as we wish.